training/requirements.txt
# GPU training dependencies for RTX 6000 (96GB VRAM)
torch>=2.1.0
transformers>=4.45.0
trl>=0.12.0
peft>=0.13.0
bitsandbytes>=0.44.0
accelerate>=1.0.0
datasets>=3.0.0
pillow>=10.0.0
opencv-python-headless>=4.8.0
mediapipe>=0.10.0
jiwer>=3.0.0
